% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{app_ui}
\alias{app_ui}
\title{User Interface for D3 Shiny App}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
app_ui
}
\description{
User Interface for D3 Shiny App
}
\keyword{datasets}
