% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gold_viz.R
\name{plot_gold_viz_bar}
\alias{plot_gold_viz_bar}
\title{Generate the Gold Visualization}
\usage{
plot_gold_viz_bar(jGold, directory_path = "D3", width = NULL)
}
\description{
Generate the Gold Visualization
}
